+++ src/auth/dtos/user-info.dto.ts
