+++ src/repositories/preference.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource, Repository } from 'typeorm';
-import { UserPreferenceDto } from 'src/user/dtos/first-login-dto';
+import { UserPreferenceDto } from 'src/user/dtos/first-login.dto';
 import { Preference } from 'src/entities/preference.entity';
 import { User } from 'src/entities/user.entity';
