+++ src/user/user.module.ts
@@ -6,8 +6,8 @@ import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { PreferenceRepository } from 'src/repositories/preference.repository';
 import { UserRepository } from 'src/repositories/user.repository';
-import { JwtAccessStrategy } from 'src/auth/strategies/jwt-access-strategy';
-import { JwtAccessNicknameCheckStrategy } from 'src/auth/strategies/jwt-access-nickname-check-strategy';
+import { JwtAccessStrategy } from 'src/auth/strategies/jwt-access.strategy';
+import { JwtAccessNicknameCheckStrategy } from 'src/auth/strategies/jwt-access-nickname-check.strategy';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, Preference])],
